move tiering tests
