+++ ydb/core/tx/tiering/ut/test.json
@@ -1,0 +1,2 @@
+{"key": "1", "value": "trololo"}
+{"key": "2", "value": "hello world"}
